+++ all-hallows-steve/common/src/main/java/com/chimericdream/allhallowssteve/AllHallowsSteveMod.java
@@ -1,5 +1,7 @@
 package com.chimericdream.allhallowssteve;
 
+import com.chimericdream.allhallowssteve.block.ModBlocks;
+import com.chimericdream.allhallowssteve.component.type.AllHallowsSteveComponentTypes;
 import com.chimericdream.lib.registries.ModRegistryHelper;
 import com.google.common.base.Suppliers;
 import dev.architectury.registry.registries.RegistrarManager;
@@ -18,5 +20,7 @@ public final class AllHallowsSteveMod {
         MANAGER = Suppliers.memoize(() -> RegistrarManager.get(ModInfo.MOD_ID));
 
         REGISTRY_HELPER.init();
+        ModBlocks.init();
+        AllHallowsSteveComponentTypes.init();
     }
 }